requirements.txt
selenium~=4.26.1
openpyxl~=3.1.5
PyGithub~=2.4.0
pytest~=8.3.3
outcome~=1.3.0.post0
streamlit~=1.38.0
github~=1.2.7
speedtest-cli
fpdf~=1.7.2
matplotlib~=3.9.2
pyjavaproperties~=0.7